requirements.txt
python-binance==1.0.19
aiohttp==3.9.1
websockets==12.0
numpy==1.24.3
pandas==2.0.3
pandas-ta==0.3.14b0
pydantic==2.5.0
structlog==23.2.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3
apscheduler==3.10.4
matrix-nio==0.21.2
openai==1.3.7
anthropic==0.7.7
tavily-python==0.3.3
requests==2.31.0
asyncio-mqtt==0.16.1
ccxt==4.1.77
ta-lib==0.4.25
scikit-learn==1.3.2
plotly==5.17.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1